fixing errors after merge, now it builds
