+++ main.cpp
@@ -34,13 +34,40 @@ int main(int argc, char ** argv) //!OCLINT tests may be long
     assert(1 == 2);
 #endif
 
-    simulation s = create_simulation(results);
-    observer o;
-    exec(s, o) ;
 
-    save_json(o,
-              convert_arc_to_string(s.get_params().i_p.net_par.net_arc) +
-              "_" + std::to_string(s.get_params().s_p.seed) + ".json");
+    auto mut_type = convert_ind_args(results).m_mutation_type;
+
+    if(mut_type == mutation_type::weights)
+    {
+
+        observer<mutation_type::weights> o;
+        auto s = create_simulation<mutation_type::weights>(results);
+        exec<mutation_type::weights>(s, o) ;
+        save_json(o,
+                  convert_arc_to_string(o.get_params().i_p.net_par.net_arc) +
+                  "_" + std::to_string(o.get_params().s_p.seed) + ".json");
+    }
+    else if (mut_type == mutation_type::activation) {
+
+        observer<mutation_type::activation> o;
+        auto s = create_simulation<mutation_type::activation>(results);
+        exec<mutation_type::activation>(s, o) ;
+        save_json(o,
+                  convert_arc_to_string(o.get_params().i_p.net_par.net_arc) +
+                  "_" + std::to_string(o.get_params().s_p.seed) + ".json");
+    }
+    else if (mut_type == mutation_type::weights_and_activation) {
+        observer<mutation_type::weights_and_activation> o;
+        auto s = create_simulation<mutation_type::weights_and_activation>(results);
+        exec<mutation_type::weights_and_activation>(s, o) ;
+        save_json(o,
+                  convert_arc_to_string(o.get_params().i_p.net_par.net_arc) +
+                  "_" + std::to_string(o.get_params().s_p.seed) + ".json");
+    }
+    else
+    {
+        throw std::runtime_error{"unknown mutation type"};
+    }
 
     return 0;
 }
