+++ parser.cpp
@@ -3,20 +3,6 @@
 #include <cassert>
 #include <map>
 
-simulation create_simulation(const cxxopts::ParseResult& parameters)
-{
-  auto env = convert_env_args(parameters);
-  auto ind = convert_ind_args(parameters);
-  auto pop = convert_pop_args(parameters);
-  auto sim = convert_sim_args(parameters);
-
-  all_params params{
-      env, ind, pop, sim
-  };
-
-  simulation s{params};
-  return s;
-}
